Pull out byte stream
